requirements.txt
PyQt5==5.15.10
mysql-connector-python==9.1.0